ai_chatbot/requirements.txt
google-generativeai>=0.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
